requirements.txt
fastapi
uvicorn
python-multipart
cloudinary>=1.25.0
jinja2
python-dotenv
